fix attention mask bugs

+++ cllm/utils.py
@@ -210,7 +210,7 @@ def _prepare_decoder_attention_mask(
         combined_attention_mask = _make_causal_mask(
             input_shape,
             # inputs_embeds.dtype,
-            torch.float32,  # [MODIFIED] force to cast to float32
+            torch.bfloat16,  # [MODIFIED] force to cast to float32
             device=inputs_embeds.device,
             past_key_values_length=past_key_values_length,
         )
